# openapi.yaml
openapi: "3.0.3"
info:
  title: Zendesk Ticket Mirror API
  version: "1.0.0"
  description: Read-only API serving mirrored Zendesk ticket data.
paths:
...
  /v1/organizations:
    get:
      operationId: listOrganizations
      summary: List organizations with pagination
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationList"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/organizations/{organizationId}:
    get:
      operationId: getOrganization
      summary: Get a single organization
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/groups:
    get:
      operationId: listGroups
      summary: List groups with pagination
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupList"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/groups/{groupId}:
    get:
      operationId: getGroup
      summary: Get a single group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Organization:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          nullable: true
        shared_tickets:
          type: boolean
          nullable: true
        shared_comments:
          type: boolean
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        details:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        group_id:
          type: integer
          format: int64
          nullable: true
    OrganizationList:
      type: object
      required: [organizations, total, limit, offset]
      properties:
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    Group:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        is_public:
          type: boolean
          nullable: true
        default:
          type: boolean
          nullable: true
        deleted:
          type: boolean
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
    GroupList:
      type: object
      required: [groups, total, limit, offset]
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
...
security:
  - {}
  - ApiKeyAuth: []